# specs/001-video-transcription-summary/contracts/api.yaml
openapi: 3.1.0
info:
  title: YTSummarizer API
  version: 1.0.0
  description: |
    Video transcription and summarization API for YouTube and public video URLs.
    
    ## Authentication
    All endpoints require JWT bearer token authentication (except /auth endpoints).
    
    ## Rate Limiting
    - API calls: 100 requests/minute per user
    - Video submissions: 10 videos/hour per user
  contact:
    name: YTSummarizer Support
    email: support@ytsummarizer.example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.ytsummarizer.com/v1
    description: Production
...
paths:
...
  /videos:
    get:
      tags: [videos]
      summary: Get user's video history
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [submitted, extracting, extracted, transcribing, transcribed, summarizing, completed, failed]
      responses:
        '200':
          description: Video list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/VideoResponse'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [videos]
      summary: Submit video URL for transcription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  example: https://www.youtube.com/watch?v=dQw4w9WgXcQ
                language:
                  type: string
                  example: en
                  description: Optional language hint (ISO 639-1 code)
      responses:
        '202':
          description: Video accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoResponse'
        '400':
          description: Invalid URL or unsupported platform
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error_code: VID-001
                message: Invalid video URL format
                detail: URL must start with http:// or https://
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /videos/{video_id}:
    get:
      tags: [videos]
      summary: Get video details and status
      security:
        - bearerAuth: []
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VideoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        platform:
          type: string
          enum: [youtube, vimeo, direct, other]
        title:
          type: string
          nullable: true
        duration_seconds:
          type: integer
          nullable: true
        thumbnail_url:
          type: string
          nullable: true
        status:
          type: string
          enum: [submitted, extracting, extracted, transcribing, transcribed, summarizing, completed, failed]
        error_code:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    VideoDetailResponse:
      allOf:
        - $ref: '#/components/schemas/VideoResponse'
        - type: object
          properties:
            has_transcription:
              type: boolean
            has_summary:
              type: boolean
            metadata:
              type: object
              additionalProperties: true
...
    Error:
      type: object
      properties:
        error_code:
          type: string
          example: VID-001
        message:
          type: string
          example: Invalid video URL format
        detail:
          type: string
          nullable: true
          example: URL must start with http:// or https://
        timestamp:
          type: string
          format: date-time
  responses:
...
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'